Fix DtaHashPwd( ) to properly encode a zero length password
When a user is activated in single-user mode, it is assigned a 0 length password by default.  To change the password, a session must be established using this 0 length password.  Per TCG Storage Architecture Core Specification, version 2.01, revision 1.00, section 3.2.2.3.1.2, a zero length byte string is encoded as 0xA0.  Fix DtaHasPwd( ) to check for a zero length string and encode it as 0xA0 instead of 0xD0 0x00.